일러스트 오류 + 경로 수정, 푸터 경로 설정 완료

+++ src/main/java/com/team1/stelling/domain/repository/IllustRepository.java
@@ -22,7 +22,7 @@ public interface IllustRepository extends JpaRepository<IllustVO, Long> {
     Page<IllustVO> findByIllustHashTagContaining(String keyword, Pageable pageable);
 
     @Query(value = "SELECT SUM(ILLUST_LIKE) FROM TBL_ILLUST WHERE USER_NUMBER = ?", nativeQuery = true)
-    int findByIllustLikeTotal(Long userNumber);
+    Long findByIllustLikeTotal(Long userNumber);
 
     @Modifying
     @Query(value = "update TBL_Illust set ILLUST_VIEWCOUNT = ILLUST_VIEWCOUNT + 1 where ILLUST_NUMBER = ?", nativeQuery = true)
